Add VM sections module
Move encode_sections() to own module
